+++ src/components/pages/NotificationCenter/NotificationSearch.tsx
@@ -18,32 +18,19 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { createAction } from '@reduxjs/toolkit'
-import type { PageNotificationsProps } from '@catena-x/portal-shared-components'
-import { type NotificationFetchType, name } from './types'
+import DebouncedSearchInput from 'components/shared/basic/Input/DebouncedSearchInput'
+import { useDispatch } from 'react-redux'
+import { setSearch } from 'features/notification/slice'
 
-const setNotification = createAction(
-  `${name}/control/setNotification`,
-  function update(notification: PageNotificationsProps) {
-    return {
-      payload: {
-        notification,
-      },
-    }
-  }
-)
+export default function NotificationSearch() {
+  const dispatch = useDispatch()
 
-const resetNotification = createAction(`${name}/resetNotification`)
-
-const resetInitialNotificationState = createAction(
-  `${name}/control/resetInitialNotificationState`,
-  function update(initialNotificationState: NotificationFetchType) {
-    return {
-      payload: {
-        initialNotificationState,
-      },
-    }
-  }
-)
-
-export { setNotification, resetNotification, resetInitialNotificationState }
+  return (
+    <div className="searchContainer">
+      <DebouncedSearchInput
+        debounceTime={500}
+        onSearch={(expr: string) => dispatch(setSearch(expr))}
+      />
+    </div>
+  )
+}